requirements.txt
openai
langchain
biopython
streamlit
python-dotenv
tqdm